refactor(suggestions): collect the 0.8.0 dividend — delete the availability probe
react-built-in-ai 0.8.0 surfaces "downloadable" as a lifecycle status,
so the app-side probe, its staleness keying, the idle disambiguation,
and the dead MissingUserActivationError branch all delete. The engine
view loses its only stateful dependency and useEngineView collapses
into the pure deriveEngineView — interpretation is now a render-time
mapping, no hook required.
