+++ package.json
@@ -36,7 +36,7 @@
     "@mui/material": "^9.0.1",
     "@mui/stylis-plugin-rtl": "^9.0.1",
     "@shayc/open-board-format": "^0.2.0",
-    "@shayc/react-built-in-ai": "^0.7.0",
+    "@shayc/react-built-in-ai": "^0.8.0",
     "idb": "^8.0.3",
     "mrmime": "^2.0.1",
     "react": "^19.2.7",
